+++ src/JsonDataFilter/Models/CountryCurrencyCode.cs
@@ -1,0 +1,9 @@
+namespace JsonDataFilter.Models
+{
+    public class CountryCurrencyCode
+    {
+        public string Country { get; set; }
+        public string CountryCode { get; set; }
+        public string CurrencyCode { get; set; }
+    }
+}
