+++ src/JsonDataFilter/Models/CountryData.cs
@@ -1,0 +1,7 @@
+namespace JsonDataFilter.Models
+{
+    public class CountryData
+    {
+        public string Country { get; set; }
+    }
+}
