+++ src/JsonDataFilter/Models/Currency.cs
@@ -1,0 +1,9 @@
+namespace JsonDataFilter.Models
+{
+    public class Currency
+    {
+        public string CountryCode { get; set; }
+        public string CountryName { get; set; }
+        public string CurrencyCode { get; set; } 
+    }
+}
