+++ src/JsonDataFilter/Startup.cs
@@ -24,6 +24,7 @@ namespace JsonDataFilter
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddRouting(options=>options.LowercaseUrls=true);
             services.AddControllers();
         }
